Added new event handlers and removed guess code
Parts of the code relating to generating and guessing the word have
been removed, as this will only suggest possible words.

There is now a separate event handler to manage the stage during which
the result tiles are edited. These are clicked on to change thier color/
state, and all must be edited before pressing enter again.

A new HTML element for the suggestion has been added. Here the suggested
word will be placed, though now it is a default text only